+++ tests/testthat/test-riskFormulations.R
@@ -1,11 +1,10 @@
-context("test-riskFormulations.R")
-library(riskParityPortfolio)
-library(testthat)
+context("Formulations")
 
 # generate a random Sigma and mu
 set.seed(123)
 N <- 5
-V <- matrix(rnorm(N^2), N, N)
+T <- 1000
+V <- matrix(rnorm(N * T), T, N)
 Sigma <- cov(V)
 
 test_that("sca, alabama, and slsq give similar results in low dimensional problems", {
@@ -14,10 +13,10 @@ test_that("sca, alabama, and slsq give similar results in low dimensional proble
                          "rc-over-sd vs b-times-sd", "rc vs b-times-var",
                          "rc vs theta", "rc-over-b vs theta")
   for(formulation in formulations_list) {
-    rpp_sca <- riskParityPortfolio(Sigma, method = "sca", formulation = formulation)
-    rpp_alabama <- riskParityPortfolio(Sigma, method = "alabama", formulation = formulation)
-    rpp_slsqp <- riskParityPortfolio(Sigma, method = "slsqp", formulation = formulation)
-    expect_that(all(abs(rpp_sca$w - rpp_alabama$w) < 1e-3), is_true())
-    expect_that(all(abs(rpp_sca$w - rpp_slsqp$w) < 1e-3), is_true())
+    suppressWarnings(rpp_sca <- riskParityPortfolio(Sigma, method = "sca", formulation = formulation))
+    suppressWarnings(rpp_alabama <- riskParityPortfolio(Sigma, method = "alabama", formulation = formulation))
+    suppressWarnings(rpp_slsqp <- riskParityPortfolio(Sigma, method = "slsqp", formulation = formulation))
+    expect_equal(rpp_sca$w, rpp_alabama$w, tolerance = 1e-3)
+    expect_equal(rpp_sca$w, rpp_slsqp$w, tolerance = 1e-3)
   }
 })