+++ NEWS.md
@@ -1,3 +1,11 @@
-# riskParityPortfolio 0.1.0: 2018-12-15
+## Changes in riskParityPortfolio version 0.1.1 (2019-01-07)
 
-* Initial release version 0.1.0 is on CRAN.
+* Revised vignette (fix name issue and include new section on algorithm description).
+* Revise the error control of riskParityPortfolio().
+* Check feasibility in riskParityPortfolio().
+* Improved tests.
+
+
+## Changes in riskParityPortfolio version 0.1.0 (2018-12-15)
+
+* Initial release is on CRAN.